change PasswordRequest use int instead of long

+++ business/account/account-api/src/main/java/study/daydayup/wolf/business/account/api/dto/request/PasswordRequest.java
@@ -28,9 +28,8 @@ public class PasswordRequest implements Request {
     private String token;
     private String scope;
 
-
-    private long expiredIn;
-    private long refreshExpiredIn;
+    private int expiredIn;
+    private int refreshExpiredIn;
 
     private AuthEnv env;
 }
